refactor(mcp): split main.rs into modules
